refactor: 파일명 통일 및 절대경로로 수정
