+++ components/ticket/MobileTicketList.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react';
 import WidePlaylistItem from './WidePlaylistItem';
-import { shuffleArray } from '../../utils/shuffleArray';
+import { shuffleArray } from '../../utils/shuffleArrayUtils';
 
 const MobileTicketList = ({
   tickets,
